# BACKEND/API/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "GPAC REST API",
    "description": "Desarrollo de una API REST para una aplicación de gestión de tareas que ayude a los usuarios a organizar y administrar sus actividades diarias. La aplicación debe permitir a los usuarios crear, seguir y completar tareas, establecer fechas límite, asignar prioridades y añadir notas según sea necesario. Además, los usuarios pueden categorizar las tareas en proyectos o etiquetas para una mejor organización. La aplicación debe ofrecer funcionalidades de búsqueda y filtrado, permitiendo a los usuarios buscar tareas específicas y aplicar filtros según criterios como fecha, prioridad o proyecto."
  },
  "host": "localhost:3000",
  "basePath": "/backend",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/misc/status/{code}": {
      "get": {
        "description": "Method used to get a status title depending on code provided. The parameter 'code' is a number between 1 to 3 and is required.",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "type": "number"          },
          {
            "name": "x-access-token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/misc/priority/{code}": {
      "get": {
        "description": "Method used to get a priority title depending on code provided. The parameter 'code' is a number between 0 to 2 and is required.",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "type": "number"          },
          {
            "name": "x-access-token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}